refactor: rename `nerdfix::util` to `nerdfix::utils`
